inotify_read_event: wait for events using select(2)
There was some problem with read() returning EINVAL for file-related
events on directory watches; this commit fixes that by

  * using the `select` system call to wait for data to become available
    for reading on the inotify FD, and by
  * using the `FIONREAD` ioctl to determine exactly how much data needs
    to be read from the file descriptor so we can read it all at once
    and process events more efficiently.

As a side-effect, we also get a timeout parameter on `inotify_read_event`
thanks to the fact that `select` implements one that is easily supported
in wrapper code.

Note that this commit slightly changes the semantics of the Python
wrapper function in question: previously it was guaranteed that the
user-supplied callback would be called only once per invocation of this
wrapper function (which is no longer the case).  The user callback will
now be called once for each item in the first "group" of events
available for reading after the wrapper function is invoked.