add support for older implementations of `inotify`
On certain supercomputing clusters, the installed version of inotify.h
predates the introduction of `inotify_init1` (which initially allowed
the programmer to specify that "close-on-execute" [IN_CLOEXEC] should be
set on the inotify file descriptor) and later additions to that file.

This commit adds support for those systems by working around the
missing interfaces to the maximum extent possible.  When needed, it

  * manually sets descriptor flags passed to the Python wrapper _after_
    creating the inotify file descriptor, and
  * skips definition of module constants that would request behavior not
    easily emulated.

This commit does *not* add full support for changes to the inotify
interface since its introduction; only those changes I need to support --
between July 2008 and today -- are handled.